requirements.txt
fastapi 
pymongo 
uvicorn 
python-dotenv
bcrypt 
pyjwt
google-genai
fastapi-sse
uuid